+++ lib/view_component/collection.rb
@@ -16,7 +16,7 @@ module ViewComponent
         content = component.new(**component_options(item, iterator)).render_in(view_context, &block)
         iterator.iterate!
         content
-      end.join.html_safe
+      end.join.html_safe # rubocop:disable Rails/OutputSafety
     end
 
     private
@@ -38,6 +38,7 @@ module ViewComponent
     def component_options(item, iterator)
       item_options = { component.collection_parameter => item }
       item_options[component.collection_counter_parameter] = iterator.index + 1 if component.counter_argument_present?
+      item_options[component.collection_iteration_parameter] = iterator if component.iteration_argument_present?
 
       @options.merge(item_options)
     end