+++ lib/view_component/slotable.rb
@@ -123,6 +123,7 @@ module ViewComponent
       slot_instance = args.present? ? slot_class.new(**args) : slot_class.new
 
       # Capture block and assign to slot_instance#content
+      # rubocop:disable Rails/OutputSafety
       slot_instance.content = view_context.capture(&block).to_s.strip.html_safe if block_given?
 
       if slot[:collection]