+++ test/view_component/view_component_test.rb
@@ -163,6 +163,12 @@ class ViewComponentTest < ViewComponent::TestCase
     ActionController::Base.allow_forgery_protection = old_value
   end
 
+  def test_renders_component_with_variant_method
+    render_inline(VariantsComponent.new.with_variant(:phone))
+
+    assert_text("Phone")
+  end
+
   def test_renders_component_with_variant
     with_variant :phone do
       render_inline(VariantsComponent.new)
@@ -266,6 +272,7 @@ class ViewComponentTest < ViewComponent::TestCase
 
   def test_renders_content_areas_template_can_wrap_render_arguments
     render_inline(ContentAreasComponent.new(title: "Hello!", footer: "Bye!")) do |component|
+      # rubocop:disable Rails/OutputSafety
       component.with(:title) { "<strong>#{component.title}</strong>".html_safe }
       component.with(:body) { "Have a nice day." }
     end
@@ -563,8 +570,8 @@ class ViewComponentTest < ViewComponent::TestCase
     render_inline(ProductComponent.with_collection(products, notice: "On sale"))
 
     assert_selector("h1", text: "Product", count: 2)
-    assert_selector("h2", text: "Radio clock")
-    assert_selector("h2", text: "Mints")
+    assert_selector("h2.first", text: "Radio clock")
+    assert_selector("h2:not(.first)", text: "Mints")
     assert_selector("p", text: "On sale", count: 2)
     assert_selector("p", text: "Radio clock counter: 1")
     assert_selector("p", text: "Mints counter: 2")
@@ -589,6 +596,20 @@ class ViewComponentTest < ViewComponent::TestCase
     assert_selector("figcaption", text: "Photo.1 - Yellow flowers")
 
     assert_selector("figure[data-index=1]", { count: 1 })
+    assert_selector("figcaption", text: "Photo.2 - Mountains at sunset")
+  end
+
+  def test_render_collection_custom_collection_parameter_name_iteration
+    photos = [
+      OpenStruct.new(title: "Flowers", caption: "Yellow flowers", url: "https://example.com/flowers.jpg"),
+      OpenStruct.new(title: "Mountains", caption: "Mountains at sunset", url: "https://example.com/mountains.jpg")
+    ]
+    render_inline(CollectionIterationComponent.with_collection(photos))
+
+    assert_selector("figure.first[data-index=0]", { count: 1 })
+    assert_selector("figcaption", text: "Photo.1 - Yellow flowers")
+
+    assert_selector("figure[data-index=1]:not(.first)", { count: 1 })
     assert_selector("figcaption", text: "Photo.2 - Mountains at sunset")
   end